Crate tauri_codegen

Modules

embedded_assets

Structs

Config

The Tauri configuration object. It is read from a file where you can define your frontend assets, configure the bundler, enable the app updater, define a system tray, enable APIs via the allowlist and more.

ContextData

Necessary data needed by context_codegen to generate code for a Tauri application context.

Enums

CodegenConfigError

Represents all the errors that can happen while reading the config during codegen.

ConfigError

Represents all the errors that can happen while reading the config.

Functions

context_codegen

Build a tauri::Context for including in application code.

get_config

Get the Config from the TAURI_CONFIG environmental variable, or read from the passed path.