Module tauri_codegen::embedded_assets

Structs

AssetOptions

Options used to embed assets.

CspHashes

Holds all hashes that we will apply on the CSP tag/header.

EmbeddedAssets

Represent a directory of assets that are compressed and embedded.

EmbeddedAssetsInput

Enums

EmbeddedAssetsError

All possible errors while reading and compressing an EmbeddedAssets directory

Type Aliases

EmbeddedAssetsResult