Struct tauri_codegen::embedded_assets::EmbeddedAssets

pub struct EmbeddedAssets { /* private fields */ }

Represent a directory of assets that are compressed and embedded.

This is the compile time generation of [tauri_utils::assets::Assets] from a directory. Assets from the directory are added as compiler dependencies by dummy including the original, uncompressed assets.

The assets are compressed during this runtime, and can only be represented as a TokenStream through ToTokens. The generated code is meant to be injected into an application to include the compressed assets in that application’s binary.

Implementations

impl EmbeddedAssets

pub fn new( input: impl Into<EmbeddedAssetsInput>, options: &AssetOptions, map: impl FnMut(&AssetKey, &Path, &mut Vec<u8>, &mut CspHashes) -> Result<(), EmbeddedAssetsError>, ) -> Result<Self, EmbeddedAssetsError>

Compress a collection of files and directories, ready to be generated into Assets.